% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_gender.R
\name{assign_gender}
\alias{assign_gender}
\title{Assign gender by first name}
\usage{
assign_gender(data_df, first_name_col)
}
\arguments{
\item{data_df, }{input dataframe containing the first name}

\item{first_name_col, }{first name column's name to assign gender to}
}
\value{
The input data frame with the gender column:

         gender - assigned gender (F/M/U)
}
\description{
This function use the data source based on combined US/UK
censor data to assign gender based on first name.
}
\examples{
gender <- assign_gender(authors, "first_name")

}
