% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{authors}
\alias{authors}
\title{names dataset}
\format{
a data frame with 1000 rows of four variables:
\describe{
  \item{first_name}{first name}
  \item{last_name}{last lame}
  \item{country_code}{country}
  \item{publication_years}{publication year}
}
}
\usage{
authors
}
\description{
This data sets contains all the name fro UK and US social security
}
\keyword{datasets}
