% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/balloon_plot.R
\name{balloon_plot}
\alias{balloon_plot}
\title{Function to create the balloon plot for gender first name}
\usage{
balloon_plot(data_df, gender_var, cutoff)
}
\arguments{
\item{data_df, }{data frame containing `first name` and `gender` columns from 
\code{\link{assign_gender}}}

\item{gender_var, }{gender possible values are F for female, M for male and U 
for unknown}

\item{cutoff, }{numerical value indicating where to cut the counting data}
}
\value{
The output is a gg object from ggplot2 which shows the most frequent
names as a balloon plot.
}
\description{
Function to create the balloon plot for gender first name
}
\examples{
gender <- assign_gender(authors, "first_name")
bp <- balloon_plot(gender, "M", cutoff = 5)
}
