% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bar_chart.R
\name{bar_chart}
\alias{bar_chart}
\title{Function to plot the barchart of the total number by gender}
\usage{
bar_chart(data_df, x_title, y_title)
}
\arguments{
\item{data_df, }{dataframe generated from the function total_df}

\item{x_title, }{title for x axis.}

\item{y_title, }{title for y axis}
}
\value{
a barchart as ggplot2 object.
}
\description{
Function to plot the barchart of the total number by gender
}
