% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseline.R
\name{baseline}
\alias{baseline}
\title{Calculate the female baseline}
\usage{
baseline(data_df, gender_col)
}
\arguments{
\item{data_df, }{dataframe containing the gender column.}

\item{gender_col, }{the name of the column containing the gender information.}
}
\value{
The function returns a numeric vector containing the baseline values
}
\description{
\code{baseline} calculate the female baseline giving a dataframe
containing the gender information.
}
\examples{
## df is the dataframe in output from the function assign_gender
df <- data.frame(first_name = c("anna", "john", "ernest", "colin", "aileen"),
                 gender = c("F", "M",  "M", "M", "F"),
                 stringsAsFactors = FALSE)
baseline <- baseline(df, gender_col = "gender")
}
