% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bullet_line_chart.R
\name{bullet_line_chart}
\alias{bullet_line_chart}
\title{Create a Bullet chart together with a line chart
to compare different baseline for gender analysis}
\usage{
bullet_line_chart(
  data_df,
  baseline_female,
  x_title,
  y_title,
  baseline_label,
  total_number_df,
  var_name,
  c,
  ysectitle,
  line_label
)
}
\arguments{
\item{data_df, }{dataframe in output from the function \code{\link{percent_df}}}

\item{baseline_female, }{numeric vector containing the baseline for each level.}

\item{x_title, }{label for x axis}

\item{y_title, }{label for y axis}

\item{baseline_label, }{label used to define the baseline name.}

\item{total_number_df, }{data frame containing the total number of submissions.}

\item{var_name, }{variable from total_number_df to plot.}

\item{c, }{factor of conversion for second y-axis}

\item{ysectitle, }{title of the second y-axis}

\item{line_label, }{label used to define the line chart.
with the corresponding baseline for a years period.}
}
\value{
This function create a bar chart containing the percentage of submission
}
\description{
Create a Bullet chart together with a line chart
to compare different baseline for gender analysis
}
\examples{
\dontrun{
bullet_line_chart(percent_df, baseline, "Submissions (\%)", "Months",
 "Women baseline", total_number_df, var_name, c, ysectitle, line_label)
}
}
