% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gender_names}
\alias{gender_names}
\title{Gender names dataset}
\format{
a data frame of two variables:
\describe{
  \item{Name}{First name}
  \item{UKUS_Gender}{Gender of the first name}
}
}
\usage{
gender_names
}
\description{
This data sets contains all the name fro UK and US social security
}
\keyword{datasets}
