% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/percent_df.R
\name{percent_df}
\alias{percent_df}
\title{Create a dataframe that will be the input to generate stacked bar chart and
bullet chart that show percentage to compare proportions among gender.}
\usage{
percent_df(data_df)
}
\arguments{
\item{data_df, }{dataframe containing level, lower_CI, upper_CI,
significance and female and male percentages}
}
\value{
dataframe with the columns x_values, y_values, gender, labels
}
\description{
Create a dataframe that will be the input to generate stacked bar chart and
bullet chart that show percentage to compare proportions among gender.
}
\examples{
\dontrun{
percent_df <-percent_df(data_df)
}
}
