% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reshape_for_binomials.R
\name{reshape_for_binomials}
\alias{reshape_for_binomials}
\title{Reshape the dataframe to apply binomial calculations.}
\usage{
reshape_for_binomials(data_df, gendercol, level)
}
\arguments{
\item{data_df, }{dataframe containing the columns gender and counts}

\item{gendercol, }{the name of the column containing the gender values.}

\item{level, }{variable to compare for the baseline.}
}
\description{
reshape dataframe from long format to wide format.
}
