% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stacked_bar_chart.R
\name{stacked_bar_chart}
\alias{stacked_bar_chart}
\title{Create a stacked bar chart with significance bars to compare with the
female baseline for gender analysis.}
\usage{
stacked_bar_chart(data_df, baseline_female, x_label, y_label, baseline_label)
}
\arguments{
\item{data_df, }{is the output dataframe from \code{\link{percent_df}}}

\item{baseline_female, }{female baseline in percentage from \code{\link{baseline}}}

\item{x_label, }{label for x axis}

\item{y_label, }{label for y axis}

\item{baseline_label, }{label used to define the baseline name.}
}
\value{
This function create a bar chart containing the percentage of
submission with the corresponding baseline.
}
\description{
Create a stacked bar chart with significance bars to compare with the
female baseline for gender analysis.
}
