% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_gd.R
\name{theme_gd}
\alias{theme_gd}
\title{gender diversity theme for ggplot2}
\usage{
theme_gd()
}
\description{
gender diversity theme for ggplot2
}
\examples{
\dontrun{
 gender_bar_chart(data_df) + theme_gd
}
}
