% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/balloon_plot.R
\name{balloon_plot}
\alias{balloon_plot}
\title{Function to create the balloon plot for gender first name}
\usage{
balloon_plot(data_df, first_name, gender, cutoff)
}
\arguments{
\item{data_df, }{data frame containing `first name` and `gender` columns}

\item{first_name, }{first name column's name}

\item{gender, }{gender possible values are F for female, M for male and U 
for unknown}

\item{cutoff, }{numerical value indicating where to cut the count data}
}
\value{
The output is a gg object from ggplot2 which shows the most frequent
names as a balloon plot.
}
\description{
Function to create the balloon plot for gender first name
}
\examples{
gender <- assign_gender(authors, "first_name")
bp <- balloon_plot(gender, "first_name", "gender", cut = 5)
}
