% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/balloon_plot.R
\name{balloon_plot}
\alias{balloon_plot}
\title{Function to create the balloon plot for gender first name}
\usage{
balloon_plot(data_df, first_name, gender)
}
\arguments{
\item{data_df, }{data frame containing the submission's data}

\item{first_name, }{first name column's name}

\item{gender, }{gender column name}
}
\description{
Function to create the balloon plot for gender first name
}
