% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bar_chart.R
\name{bar_chart}
\alias{bar_chart}
\title{Function to create a bar chart of the total number by gender}
\usage{
bar_chart(data_df, x_label, y_label)
}
\arguments{
\item{data_df, }{dataframe from \code{\link{total_gender_df}}}

\item{x_label, }{label for x axis.}

\item{y_label, }{label for y axis.}
}
\value{
A bar chart as ggplot2 object showing on the y axis the
 total number per gender and on the x axis the level previously defined in
 \code{\link{total_gender_df}}.
}
\description{
Function to create a bar chart of the total number by gender
}
