% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bullet_line_chart.R
\name{bullet_line_chart}
\alias{bullet_line_chart}
\title{Create a Bullet chart together with a line chart
to compare different baseline for gender analysis}
\usage{
bullet_line_chart(
  data_df,
  baseline_female,
  x_title,
  y_bar_chart_title,
  baseline_label,
  line_chart_df,
  line_chart_scaling,
  y_line_chart_title,
  line_label
)
}
\arguments{
\item{data_df, }{dataframe in output from \code{\link{percent_df}}}

\item{baseline_female, }{numeric vector containing the baseline for each level}

\item{x_title, }{label for x axis for both charts}

\item{y_bar_chart_title, }{label for y axis of the bar chart}

\item{baseline_label, }{label used to define the baseline name.}

\item{line_chart_df, }{data frame containing the total number of submissions}

\item{line_chart_scaling, }{factor of conversion for second y-axis}

\item{y_line_chart_title, }{title the y-axis of the line chart}

\item{line_label, }{label used to define the line chart.}
}
\value{
The function create a bar chart containing the percentage of male and
female and the total number of submissions on the top chart.
}
\description{
Create a Bullet chart together with a line chart
to compare different baseline for gender analysis
}
