% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bullet_line_chart.R
\name{bullet_line_chart}
\alias{bullet_line_chart}
\title{Function to create a bullet chart with a line chart in the same graphical 
frame; to compare different baselines for gender analysis.}
\usage{
bullet_line_chart(
  data_df,
  baseline_female,
  x_label,
  y_bullet_chart_label,
  baseline_label,
  line_chart_df,
  line_chart_scaling,
  y_line_chart_label,
  line_label
)
}
\arguments{
\item{data_df, }{dataframe in output from \code{\link{percent_df}}}

\item{baseline_female, }{numeric vector containing the baseline for each level}

\item{x_label, }{label for x axis for both charts}

\item{y_bullet_chart_label, }{label for y axis of the bullet chart}

\item{baseline_label, }{label used to define the baseline name.}

\item{line_chart_df, }{data frame containing the total number of submissions}

\item{line_chart_scaling, }{factor of conversion for second y-axis}

\item{y_line_chart_label, }{label the y-axis of the line chart}

\item{line_label, }{label used to define the line chart.}
}
\value{
The function create a bullet chart containing the percentage of male 
and female with the corresponding baseline for the level defined in
\code{\link{percent_df}}. The total number of submissions are displayed on 
the top of the bullet chart.
}
\description{
Function to create a bullet chart with a line chart in the same graphical 
frame; to compare different baselines for gender analysis.
}
