% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stacked_bar_chart.R
\name{stacked_bar_chart}
\alias{stacked_bar_chart}
\title{Create a stacked bar chart with significance bars to compare with the
baseline for gender analysis}
\usage{
stacked_bar_chart(
  data_df,
  baseline_female_percentage,
  x_title,
  y_title,
  baseline_label
)
}
\arguments{
\item{data_df, }{is the output dataframe from the function percent_df}

\item{baseline_female_percentage, }{female baseline in percentage}

\item{x_title, }{label for x axis}

\item{y_title, }{label for y axis}

\item{baseline_label, }{label used to define the baseline name.}
}
\value{
This function create a bar chart containing the percentage of
submission with the corresponding baseline.
}
\description{
Create a stacked bar chart with significance bars to compare with the
baseline for gender analysis
}
\examples{
\dontrun{
stacked_bar_chart(percent_df, baseline,
"Submissions (\%)", "Months", "Women baseline")
}
}
