% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_gd.R
\name{theme_gd}
\alias{theme_gd}
\title{This function create a gender diversity theme for chart based on ggplot2}
\usage{
theme_gd()
}
\value{
an object of the class theme defined in ggplot2 own class system.
}
\description{
This function create a gender diversity theme for chart based on ggplot2
}
\examples{
 require(ggplot2)
 ggplot(authors, aes(x = publication_years)) + geom_bar() + theme_gd()
}
