% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/total_gender_df.R
\name{total_gender_df}
\alias{total_gender_df}
\title{Create a dataframe that will be the input to generate the bar chart of
the full amount of female and male}
\usage{
total_gender_df(data_df, level)
}
\arguments{
\item{data_df, }{dataframe from \code{\link{calculate_binom_baseline}}
containing Level, LCI, UCI,
Significance and Male and Female percentages}

\item{level, }{name of level}
}
\value{
The output is a dataframe with the columns x_values, 
total_female_male, gender, y_values. This data frame is the input to create 
the bar chart for \code{\link{bar_chart}}
}
\description{
Create a dataframe that will be the input to generate the bar chart of
the full amount of female and male
}
